code review comments and optimizations
