+++ Tests/AEPOptimizeTests/UnitTests/OptimizePublicAPITests.swift
@@ -550,194 +550,6 @@ class OptimizePublicAPITests: XCTestCase {
         wait(for: [expectation], timeout: 1)
     }
     
-    func testTrackDisplayedOffers() {
-        // setup
-        let expectation = XCTestExpectation(description: "trackDisplayedOffers should dispatch an event with expected data.")
-        expectation.assertForOverFulfill = true
-        
-        // Create test offers
-        let offerData1 = """
-        {
-            "id": "xcore:personalized-offer:1111111111111111",
-            "schema": "https://ns.adobe.com/experience/offer-management/content-component-text",
-            "data": {
-                "id": "xcore:personalized-offer:1111111111111111",
-                "type": 2,
-                "content": "Test content 1"
-            }
-        }
-        """.data(using: .utf8)!
-        
-        let offerData2 = """
-        {
-            "id": "xcore:personalized-offer:2222222222222222",
-            "schema": "https://ns.adobe.com/experience/offer-management/content-component-text",
-            "data": {
-                "id": "xcore:personalized-offer:2222222222222222",
-                "type": 2,
-                "content": "Test content 2"
-            }
-        }
-        """.data(using: .utf8)!
-        
-        let offerData3 = """
-        {
-            "id": "xcore:personalized-offer:3333333333333333",
-            "schema": "https://ns.adobe.com/experience/offer-management/content-component-text",
-            "data": {
-                "id": "xcore:personalized-offer:3333333333333333",
-                "type": 2,
-                "content": "Test content 3"
-            }
-        }
-        """.data(using: .utf8)!
-        
-        guard let offer1 = try? JSONDecoder().decode(Offer.self, from: offerData1),
-              let offer2 = try? JSONDecoder().decode(Offer.self, from: offerData2),
-              let offer3 = try? JSONDecoder().decode(Offer.self, from: offerData3) else {
-            XCTFail("Offers should be valid.")
-            return
-        }
-        
-        // Create propositions
-        let propositionData1 = """
-        {
-            "id": "de03ac85-802a-4331-a905-a57053164d35",
-            "scope": "eyJ4ZG06YWN0aXZpdHlJZCI6ImRwczpvZmZlci1hY3Rpdml0eToxYTc4OWFkYTE0ODQ1YjA2IiwieGRtOnBsYWNlbWVudElkIjoiZHBzOm9mZmVyLXBsYWNlbWVudDoxYTc4Njc0YWI1MDg1MDZjIn0=",
-            "scopeDetails": {},
-            "items": [
-                {
-                    "id": "xcore:personalized-offer:1111111111111111",
-                    "schema": "https://ns.adobe.com/experience/offer-management/content-component-text",
-                    "data": {
-                        "id": "xcore:personalized-offer:1111111111111111",
-                        "type": 2,
-                        "content": "Test content 1"
-                    }
-                },
-                {
-                    "id": "xcore:personalized-offer:2222222222222222",
-                    "schema": "https://ns.adobe.com/experience/offer-management/content-component-text",
-                    "data": {
-                        "id": "xcore:personalized-offer:2222222222222222",
-                        "type": 2,
-                        "content": "Test content 2"
-                    }
-                }
-            ]
-        }
-        """.data(using: .utf8)!
-        
-        let propositionData2 = """
-        {
-            "id": "AT:eyJhY3Rpdml0eUlkIjoiMTI1NTg5IiwiZXhwZXJpZW5jZUlkIjoiMCJ9",
-            "scope": "eyJhY3Rpdml0eUlkIjoieGNvcmU6b2ZmZXItYWN0aXZpdHk6MTExMTExMTExMTExMTExMSIsInBsYWNlbWVudElkIjoieGNvcmU6b2ZmZXItcGxhY2VtZW50OjExMTExMTExMTExMTExMTEifQ==",
-            "scopeDetails": {
-                "decisionProvider": "TGT",
-                "activity": { "id": "125589" },
-                "experience": { "id": "0" },
-                "strategies": [{"algorithmID": "0", "trafficType": "0"}]
-            },
-            "items": [
-                {
-                    "id": "xcore:personalized-offer:3333333333333333",
-                    "schema": "https://ns.adobe.com/experience/offer-management/content-component-text",
-                    "data": {
-                        "id": "xcore:personalized-offer:3333333333333333",
-                        "type": 2,
-                        "content": "Test content 3"
-                    }
-                }
-            ]
-        }
-        """.data(using: .utf8)!
-        
-        guard let proposition1 = try? JSONDecoder().decode(OptimizeProposition.self, from: propositionData1),
-              let proposition2 = try? JSONDecoder().decode(OptimizeProposition.self, from: propositionData2) else {
-            XCTFail("Propositions should be valid.")
-            return
-        }
-        
-        // Manually associate offers with propositions
-        offer1.proposition = proposition1
-        offer2.proposition = proposition1
-        offer3.proposition = proposition2
-        
-        
-        let testEvent = Event(name: "Optimize Track Propositions Request",
-                             type: "com.adobe.eventType.optimize",
-                             source: "com.adobe.eventSource.requestContent",
-                              data: nil)
-        
-        EventHub.shared.getExtensionContainer(MockExtension.self)?.registerListener(type: testEvent.type,
-                                                                                  source: testEvent.source) { event in
-            XCTAssertEqual(event.name, testEvent.name)
-            XCTAssertEqual(event.type, testEvent.type)
-            XCTAssertEqual(event.source, testEvent.source)
-            expectation.fulfill()
-        }
-        
-        // test
-        Optimize.displayed(for: [offer1, offer2, offer3])
-        
-        // verify
-        wait(for: [expectation], timeout: 2)
-    }
-    
-    func testGenerateInteractionXdm_multiplePropositions() throws {
-        guard
-            let propositionData1 = Propositions_OptimizeTests.shared.PROPOSITION_VALID.data(using: .utf8),
-            let proposition1 = try? JSONDecoder().decode(OptimizeProposition.self, from: propositionData1),
-            let propositionData2 = Propositions_OptimizeTests.shared.PROPOSITION_VALID_TARGET.data(using: .utf8),
-            let proposition2 = try? JSONDecoder().decode(OptimizeProposition.self, from: propositionData2)
-        else {
-            XCTFail("Proposition should be valid.")
-            return
-        }
-        
-        guard let propositionInteractionXdm = OptimizeTrackingUtils.generateInteractionXdm(for: [proposition1, proposition2],
-                                                                              for: OptimizeConstants.JsonValues.EE_EVENT_TYPE_PROPOSITION_DISPLAY) else {
-            XCTFail("Generated proposition interaction XDM should be valid.")
-            return
-        }
-        
-        let eventType = try XCTUnwrap(propositionInteractionXdm["eventType"] as? String)
-        XCTAssertEqual("decisioning.propositionDisplay", eventType)
-
-        let experience = try XCTUnwrap(propositionInteractionXdm["_experience"] as? [String: Any])
-        let decisioning = try XCTUnwrap(experience["decisioning"] as? [String: Any])
-        let propositionEventType = try XCTUnwrap(decisioning["propositionEventType"] as? [String: Any])
-        XCTAssertEqual(1, propositionEventType["display"] as? Int)
-        let propositionInteractionDetailsArray = try XCTUnwrap(decisioning["propositions"] as? [[String: Any]])
-        XCTAssertEqual(2, propositionInteractionDetailsArray.count)
-
-        let propositionInteractionDetails1 = try XCTUnwrap(propositionInteractionDetailsArray[0])
-        XCTAssertEqual(proposition1.id, propositionInteractionDetails1["id"] as? String)
-        XCTAssertEqual(proposition1.scope, propositionInteractionDetails1["scope"] as? String)
-
-        let scopeDetails = propositionInteractionDetails1["scopeDetails"] as? [String: Any] ?? [:]
-        XCTAssertTrue(scopeDetails == [:])
-
-        let items = try XCTUnwrap(propositionInteractionDetails1["items"] as? [[String: Any]])
-        XCTAssertEqual(1, items.count)
-
-        let item = items[0]
-        XCTAssertEqual("xcore:personalized-offer:1111111111111111", item["id"] as? String)
-        
-        let propositionInteractionDetails2 = try XCTUnwrap(propositionInteractionDetailsArray[1])
-        XCTAssertEqual(proposition2.id, propositionInteractionDetails2["id"] as? String)
-        XCTAssertEqual(proposition2.scope, propositionInteractionDetails2["scope"] as? String)
-
-        let scopeDetails2 = propositionInteractionDetails2["scopeDetails"] as? [String: Any] ?? [:]
-        XCTAssertTrue(proposition2.scopeDetails == scopeDetails2)
-
-        let items2 = try XCTUnwrap(propositionInteractionDetails2["items"] as? [[String: Any]])
-        XCTAssertEqual(1, items2.count)
-
-        let item2 = items2[0]
-        XCTAssertEqual("246315", item2["id"] as? String)
-    }
-
     func testClearCachedPropositions() {
         // setup
         let expectation = XCTestExpectation(description: "clearCachedPropositions should dispatch an event.")
